Added functionality for quicksave and quickload
