+++ Cargo.toml
@@ -10,6 +10,9 @@ edition = "2018"
 bitfield = "0.13.2"
 rand = "0.7.3"
 sdl2 = { version = "*", features = ["bundled", "static-link"] }
+serde = { version = "1.0", features = ["derive", "rc"] }
+serde-big-array = "0.3.0"
+serde_json = "1.0.57"
 
 [profile.release]
 debug = true
