+++ src/audio/noise.rs
@@ -1,10 +1,11 @@
+use serde::{Serialize, Deserialize};
 use super::envelope;
 
 const PERIOD_TABLE: [u16; 16] = [
     4, 8, 16, 32, 64, 96, 128, 160, 202, 254, 380, 508, 762, 1016, 2034, 4068
 ];
 
-#[derive(Debug)]
+#[derive(Serialize, Deserialize, Debug)]
 pub struct Noise {
     pub envelope: envelope::Envelope,
     pub length_counter: u8,
@@ -40,7 +41,7 @@ impl Noise {
         if self.is_silenced() {
             0
         } else if self.constant_volume {
-            self.envelope.decay_counter_period
+            self.envelope.volume
         } else {
             self.envelope.decay_counter
         }
@@ -52,14 +53,16 @@ impl Noise {
         }
     }
 
+    /// 0x400C
     pub fn set_volume(&mut self, data: u8) {
         self.envelope.loop_flag = (data & 0b100000) > 0;
         self.constant_volume = (data & 0b10000) > 0;
 
         let volume = data & 0b1111;
-        self.envelope.decay_counter_period = volume;
+        self.envelope.volume = volume;
     }
 
+    /// 0x400E
     pub fn set_period(&mut self, data: u8) {
         self.mode = data & 0b10000000 > 0;
         let period_index = (data & 0b1111) as usize;
