+++ src/instant/mod.rs
@@ -1,0 +1,13 @@
+use std::time::Instant;
+
+pub struct InstantWrapper {
+    pub instant: Instant
+}
+
+impl Default for InstantWrapper {
+    fn default() -> Self {
+        InstantWrapper {
+            instant: Instant::now()
+        }
+    }
+}
