+++ src/mappers/mod.rs
@@ -6,4 +6,5 @@ pub mod mapper002;
 pub mod mapper003;
 pub mod mapper004;
 pub mod mapper066;
+pub mod mapper_save_data;
 pub mod mapper_results;
