+++ src/memory_sizes.rs
@@ -1,5 +1,6 @@
 pub const KILOBYTES_1: u16 = 0x0400;
 pub const KILOBYTES_1_MASK: u16 = KILOBYTES_1 - 1;
+pub const KILOBYTES_2: u16 = 0x0800;
 pub const KILOBYTES_4: u16 = 0x1000;
 pub const KILOBYTES_4_MASK: u16 = KILOBYTES_4 - 1;
 pub const KILOBYTES_8: u16 = 0x2000;
