+++ src/ppu/name_table.rs
@@ -1,18 +1,24 @@
-use std::rc::Rc;
+use serde::{Serialize, Deserialize};
+use std::fmt::{Debug, Formatter, Result};
 use std::cell::RefCell;
+use std::rc::Rc;
 use crate::cartridge::Cartridge;
 use crate::cartridge::mirror::Mirror;
 use crate::memory_sizes::*;
 
+big_array! { BigArray; }
+
 /// A full name table is 1KB and the NES can hold 2 name tables
+#[derive(Serialize, Deserialize)]
 pub struct NameTable {
-    data: [[u8; KILOBYTES_1 as usize]; 2]
+    #[serde(with = "BigArray")]
+    data: [u8; KILOBYTES_2 as usize]
 }
 
 impl NameTable {
     pub fn new() -> Self {
         NameTable {
-            data: [[0; KILOBYTES_1 as usize]; 2]
+            data: [0; KILOBYTES_2 as usize]
         }
     }
     
@@ -23,25 +29,25 @@ impl NameTable {
                 match c.borrow().get_mirror() {
                     Mirror::Vertical => {
                         match name_table_address.masked_address {
-                            0x0000..=0x03FF => self.data[0][name_table_address.address_offset],
-                            0x0400..=0x07FF => self.data[1][name_table_address.address_offset],
-                            0x0800..=0x0BFF => self.data[0][name_table_address.address_offset],
-                            0x0C00..=0x0FFF => self.data[1][name_table_address.address_offset],
+                            0x0000..=0x03FF => self.data[name_table_address.address_offset],
+                            0x0400..=0x07FF => self.data[(KILOBYTES_1 as usize) + name_table_address.address_offset],
+                            0x0800..=0x0BFF => self.data[name_table_address.address_offset],
+                            0x0C00..=0x0FFF => self.data[(KILOBYTES_1 as usize) + name_table_address.address_offset],
                             _ => 0
                         }
                     },
                     Mirror::Horizontal => {
                         match name_table_address.masked_address {
-                            0x0000..=0x07FF => self.data[0][name_table_address.address_offset],
-                            0x0800..=0x0FFF => self.data[1][name_table_address.address_offset],
+                            0x0000..=0x07FF => self.data[name_table_address.address_offset],
+                            0x0800..=0x0FFF => self.data[(KILOBYTES_1 as usize) + name_table_address.address_offset],
                             _ => 0
                         }
                     },
                     Mirror::OneScreenLow => {
-                        self.data[0][name_table_address.address_offset]
+                        self.data[name_table_address.address_offset]
                     },
                     Mirror::OneScreenHigh => {
-                        self.data[1][name_table_address.address_offset]
+                        self.data[(KILOBYTES_1 as usize) + name_table_address.address_offset]
                     }
                     _ => 0
                 }
@@ -57,25 +63,25 @@ impl NameTable {
                 match c.borrow().get_mirror() {
                     Mirror::Vertical => {
                         match name_table_address.masked_address {
-                            0x0000..=0x03FF => self.data[0][name_table_address.address_offset] = data,
-                            0x0400..=0x07FF => self.data[1][name_table_address.address_offset] = data,
-                            0x0800..=0x0BFF => self.data[0][name_table_address.address_offset] = data,
-                            0x0C00..=0x0FFF => self.data[1][name_table_address.address_offset] = data,
+                            0x0000..=0x03FF => self.data[name_table_address.address_offset] = data,
+                            0x0400..=0x07FF => self.data[(KILOBYTES_1 as usize) + name_table_address.address_offset] = data,
+                            0x0800..=0x0BFF => self.data[name_table_address.address_offset] = data,
+                            0x0C00..=0x0FFF => self.data[(KILOBYTES_1 as usize) + name_table_address.address_offset] = data,
                             _ => ()
                         }
                     },
                     Mirror::Horizontal => {
                         match name_table_address.masked_address {
-                            0x0000..=0x07FF => self.data[0][name_table_address.address_offset] = data,
-                            0x0800..=0x0FFF => self.data[1][name_table_address.address_offset] = data,
+                            0x0000..=0x07FF => self.data[name_table_address.address_offset] = data,
+                            0x0800..=0x0FFF => self.data[(KILOBYTES_1 as usize) + name_table_address.address_offset] = data,
                             _ => ()
                         }
                     },
                     Mirror::OneScreenLow => {
-                        self.data[0][name_table_address.address_offset] = data
+                        self.data[name_table_address.address_offset] = data
                     },
                     Mirror::OneScreenHigh => {
-                        self.data[1][name_table_address.address_offset] = data
+                        self.data[(KILOBYTES_1 as usize) + name_table_address.address_offset] = data
                     }
                     _ => ()
                 }
@@ -99,5 +105,13 @@ impl NameTableAddress {
             masked_address,
             address_offset
         }
+    }
+}
+
+impl Debug for NameTable {
+    fn fmt(&self, f: &mut Formatter<'_>) -> Result {
+        f.debug_struct("NameTable")
+         .field("data_length", &self.data.len())
+         .finish()
     }
 }
